% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pothole_model.R
\docType{data}
\name{pothole_model}
\alias{pothole_model}
\title{Edmonton Pothole Prediction Model}
\format{a \code{\link{ets}} object}
\source{
https://dev.socrata.com/foundry/#/dashboard.edmonton.ca/i3wp-57z9
}
\usage{
pothole_model
}
\description{
Simple example model that predicts number of potholes filled monthly in Edmonton
}

