% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pothole_predict.R
\name{pothole_predict}
\alias{pothole_predict}
\title{Predict Potholes filled Monthly in Edmonton}
\usage{
pothole_predict(input)
}
\arguments{
\item{input}{data passed on as \code{h} to \code{\link{forecast}}}
}
\description{
Simple forecast model
}
\examples{
   mydata <- data.frame(month='2016-01-01')
   pothole_predict(mydata)
}

